[443504] Fix computation of ignored elements when finding insertion idx
When finding the insertion index of a multi-valued feature, we now make
sure that values of the target list are only ignored in the computation
of the LCS, when there is another unresolved diff element that not only
concerns the same feature and same value but also the same target
container.

Therefore, we extended and refactored the UnresolvedDiffMatching
predicate, which is used for filtering out values to be ignored, and
provide it with the additional information needed for determining the
target container (i.e., direction of merging).

Bug: 443504
Change-Id: I8b9978c1e21593db6fd092e1e87b032ea0f5f425
Signed-off-by: Philip Langer <planger@eclipsesource.com>