[450005] Adds specific handling of opaque action, behavior & expression
Opaque actions, behaviors, and expressions all have a multi-valued
language and multi-valued body attribute. Their indices have to be kept
in sync, because the language value at the index n denotes the language
of the body value at the index n. Moreover, if concurrent changes to the
same body value are applied, they should be merged using a line-based
three-way merge algorithm. If they are not mergeable with such an
algorithm, a specific conflict should be raised.

This commit contributes support for the aforementioned requirements by
introducing a new UML-specific change type, specific conflict detection,
and a specific merger.

Bug: 450005
Change-Id: I6bf5c786e72a10c56eff7f0faaedd67b77528b4b
Signed-off-by: Philip Langer <planger@eclipsesource.com>