[452147] Fix handling of diffs for interlocked one-to-one refs
Every difference describes an effect which modifies the model. In a
model with interlocked one-to-one references merging a difference can
implicitly execute the effect of other differences without being
equivalent to them. These differences are added to the
directResultingMerges of the AbstractMerger.

This patch also modifies the determination of equivalences for the
described case in the DefaultEquiEngine.

Includes testcases.

Bug: 452147
Signed-off-by: Stefan Dirix <sdirix@eclipsesource.com>
Change-Id: I8741885c45736f82e2442a32843f7d0afbaf9a67